Fix: Moved app.ts and server.ts into src/api/v1 (whoops)
