+++ src/api/v1/errors/errors.ts
@@ -1,4 +1,4 @@
-import { HTTP_STATUS } from "../../../../constants/httpConstants";
+import { HTTP_STATUS } from "../constants/httpConstants";
 
 /**
  * Base error class for application errors.
